Fix run.sh for OSX (WIP)

+++ loop/e2e-test/run.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 
 PATH_MARIONETTE="$PWD/tests/marionette/marionette"
-#PATH_FIREFOX="/Applications/Firefox.app/Contents/MacOS"
-PATH_FIREFOX="$PWD/firefox"
 PATH_INI="$PWD/caller_case.ini"
 VENV_BIN="$PWD/marionette_env/bin"
+
 
 echo
 echo "------------------------------------------------"
@@ -15,19 +14,34 @@ echo
 virtualenv marionette_env
 . $PWD/marionette_env/bin/activate 
 $VENV_BIN/pip install marionette_client six pexpect pyperclip mozdownload
-sudo apt-get install xclip
 
+echo "----------------------------------"
+echo "INSTALL FIREFOX: $OSTYPE"
+echo "----------------------------------"
 echo
-echo "------------------------------------------------"
-echo "DOWNLOAD FIREFOX"
-echo "------------------------------------------------"
-echo
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
 
-# [1]. download Firefox
-# TODO: figure out downloading: Nightly, Beta, etc.
-$VENV_BIN/mozdownload --version=latest
-rm -rf firefox
-tar xjfv *.bz2
+    echo "SET FIREFOX BIN PATH"
+    PATH_FIREFOX="$PWD/firefox"
+
+    echo "INSTALL LINUX DEPS"
+    sudo apt-get install xclip
+
+    echo "DOWNLOAD FIREFOX"
+    $VENV_BIN/mozdownload --version=latest
+
+    echo "CLEANUP"
+    rm -rf firefox
+    tar xjfv *.bz2
+
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+    echo "SET FIREFOX BIN PATH"
+    # TODO: mount new dmg, hard-coding to existing for now
+    PATH_FIREFOX="/Applications/Firefox.app/Contents/MacOS"
+else
+    echo "don't recognize OS... ABORTING!"
+    exit
+fi
 
 echo
 echo "------------------------------------------------"
@@ -35,14 +49,10 @@ echo "DOWNLOAD MARIONETTE TESTS"
 echo "------------------------------------------------"
 echo
 
-# [2]. download common.tests.zip
 $VENV_BIN/mozdownload --type=daily --extension=common.tests.zip
-# xclip needed for ubuntu
-# [4]. mkdir tests
 rm -rf tests;
 mkdir tests
 
-# [5]. unzip common.tests.zip in ./tests dir
 cp *common.tests.zip tests
 unzip  *common.tests.zip -d tests
 
@@ -54,5 +64,3 @@ echo
 
 #$VENV_BIN/python "$PATH_MARIONETTE/runtests.py" --binary="$PATH_FIREFOX/firefox-bin" --address=localhost:2828 --type=browser $PATH_INI 
 python "$PATH_MARIONETTE/runtests.py" --binary="$PATH_FIREFOX/firefox-bin" --address=localhost:2828 --type=browser $PATH_INI 
-
-
